+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
-## 3.7.0-beta - 2025-03-05
+## 3.8.0-beta - 2025-07-11
+### Added
+- YLP-2344 Patient profile information page for HCPs only
+- YLP-3589 add new estimated total insulin statistic
+
+### Changed
+- YLP-2347 Display patient monitoring parameters in the patient profile page
+- YLP-3651 Transfer the patient information displayed in the banner in the patient profile page
+
+## 3.7.0 - 2025-07-10
 ### Added
 - YLP-3151 Display Night mode events in Daily view
 
@@ -14,6 +23,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-3549 Update dashboard cards UI
 - YLP-3574 Meal bolus correction display
 - YLP-3636 Update banner to display HTML content
+- YLP-3652 Do not display INSULIN_TYPE_USED in parameters list
 
 ### Fixed
 - YLP-3290 Top banner is hiding the top part of note creation popup